+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.1.0] - 2023-05-29
+### Added
+- Added a Portal for the product-list-add-to-cart picker.
+- Added feature to configure a modal.
 
 ## [1.0.3] - 2020-04-15
 ### Fixed